feat: add testkit module
