--- e2e/src/test/java/dev/tachyonmcp/e2e/Mcp20251125TestClient.java
@@ -1,14 +1,0 @@
-/* Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors. */
-package dev.tachyonmcp.e2e;
-
-final class Mcp20251125TestClient extends TestMcpClient {
-
-    Mcp20251125TestClient(int port) {
-        super(port);
-    }
-
-    @Override
-    protected String protocolVersion() {
-        return "2025-11-25";
-    }
-}
