+++ tachyon-testkit/src/main/java/dev/tachyonmcp/testkit/package-info.java
@@ -1,0 +1,5 @@
+/* Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors. */
+@NullMarked
+package dev.tachyonmcp.testkit;
+
+import org.jspecify.annotations.NullMarked;
